Added basic web scraping functionality
Blarg. After finding out that what was on the webpage did not easily map to what comes out of the feeds, I decided to go with scraping the page.
Which meant I needed to learn about headless browsers. I didn't have the energy to figure out how to get the entire DOM accessible -- so I went with a snapshot of the accessibility API.